% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_doc.R
\docType{data}
\name{gene_set_list}
\alias{gene_set_list}
\title{GeneSet objects from ChIP-Seq experiments}
\format{a list of 1059 GeneSet objects.}
\usage{
data("gene_set_list")
}
\description{
List of 1059 GeneSets containing Entrez Gene IDs from genes assigned to
1059 ChIP-Seq experiments.
}
\keyword{datasets}
